Refactor/linear regression (#36)
* Improve documentation
Added a README.md in the subfolder in order to better understand what is happening in the code.

* Add code running section in README.md
The code is a bit complex to grok, will need to fix multiple section to make it nicer.

* Reduce filespread in the codebase
There was way too much filespread in the codebase I've shrunk it down by keeping the same
code, but moving it around.

* Refactored the load between utils and main
Some classes that were core to main were in the utils.cpp file which made things a bit too clumsy.

* Remove useless function in Dataset
I'm in the midst of refactoring the classes and I'm trimming down useless stuff until
I get a minimalistic working version.

* Improve code structure
I've shuffled the class definition around and kept things that I can reuse in other project into
math.cpp/h. This separation will help make sure that the code isn't tied to the model.

* Improved the math.cpp code
I removed the bloat in the file and increased the documentation when necessary

* Remove a read_csv standalone function
This wasn't useful and was absorbed directly into the Dataset class instead!

* Clean up the main.cpp LinearRegressionModel
I've removed some more bloat in the model as well as simplifying a lot the Dataset class

* Refactor the print_weights
It was better fitted in the Weights class. There are some improvement to do though
since the Weights class is now doing a bit too much.

* Fix the variable init bug in LinearRegressionModel
There was a bug where the internal variable where not set, it's now fixed!

* move length into private variables section

* Refactor the Weights Initialization
It was easier to just have a constructor instead of doing the excessive gymnastic!

* Add last change to linear regression

Co-authored-by: Yacine Mahdid <yacine@grad4.com>